Misc. optimizations (#117)
* Optimize Text equality comparison

* Replace clone() -> fast_clone() where possible in vm.rs

* Fix various linter warnings

* Run 'cargo clippy --fix'

* Run 'cargo fmt'

* Fix more linter warnings

* Remove clone()

* Remove unused Source::ExpStep

* Resolve prim functions at compile time

+++ crates/motoko/tests/test_vm.rs
@@ -439,6 +439,6 @@ fn test_core_eval() {
     let y = core.eval("y").expect("oops");
     assert_eq!(
         &*y,
-        &motoko::value::Value::Nat(num_bigint::BigUint::from(2 as u32))
+        &motoko::value::Value::Nat(num_bigint::BigUint::from(2_u32))
     )
 }
